test: consolidate composition coverage
